Cancel ping timeout after successful join
First, this is needed for webAO to not fall through the cracks in the timeout.

Second, putting a timeout probably was not helping in the long run, since the
TCP stack already has its own timeout which will reset the connection if we are
not receiving ACKs from the IC messages we are broadcasting.