Re-add /mute and /unmute all variants, fix "/kick" all from server
"/mute *" and "/unmute *"
Mutes or unmutes everyone in the current area ("/mute *" ignores logged in mods).

Changed the case for /kick'ing everyone from server from "**" to "@", as the former merely registered as "*" and would only kick from the area.

Also added an OOC error message when a mod attempts to kick all when they are the only user instead of throwing up an internal error.